Working on GPS cycling on first run (Issue #6).
This commit solves the issue, but forces the notification bar to always be present and always runs in foreground.  This is probably desired anyway, but I'm not closing this issue in this commit because the preference to hide the notification widget needs to be removed or another work-around to the foreground of the service needs to be implemented.